Test illustrating usage of @SuppressWarnings("null")
The @SuppressWarnings("null") should work, based on:
https://help.eclipse.org/oxygen/index.jsp?topic=%2Forg.eclipse.jdt.doc.user%2Ftasks%2Ftask-suppress_warnings.htm

However currently both on the mvn CLI and in Eclipse IDE UI Photon Milestone 2 (4.8.0M2)
this appears to be completely ignored.  Even @SuppressWarnings("all") is ignored:

Null type mismatch: required '@NonNull Object' but the provided value is null

This is, I'm guessing, because here we've intentionally configured it to
be an Error, not a Warning - which is what we would want for production code.
However, one may still want to write a test for it - and suppress it there.
Currently this appears to be impossible.